requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
email-validator==2.2.0
SQLAlchemy==2.0.36
alembic==1.13.3
PyMySQL==1.1.1
bcrypt==4.2.0
cryptography==43.0.3
jinja2==3.1.4
python-multipart==0.0.17
psutil==6.1.0
pyjwt==2.9.0
starlette==0.41.2
